backend/requirements.txt
Flask==2.0.1
Werkzeug==2.0.3
flask-cors==3.0.10
flask-socketio==5.3.0
python-nmap==0.7.1
scapy==2.5.0
networkx==3.0
matplotlib
ipaddress==1.0.23
typing_extensions==4.7.1
psutil==5.9.5
netifaces==0.11.0